Change from increment to bigIncrement

+++ database/migrations/2016_06_01_000004_create_oauth_clients_table.php
@@ -12,8 +12,8 @@ class CreateOauthClientsTable extends Migration
     public function up()
     {
         Schema::create('oauth_clients', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->index()->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->index()->nullable();
             $table->string('name');
             $table->string('secret', 100);
             $table->text('redirect');
