+++ database/migrations/2018_12_14_181700_create_password_histories_table.php
@@ -12,8 +12,8 @@ class CreatePasswordHistoriesTable extends Migration
     public function up()
     {
         Schema::create('password_histories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('password');
             $table->timestamps();
         });
